final code

+++ lib/ArduinoJson/src/ArduinoJson/Strings/FlashStringAdapter.hpp
@@ -1,10 +1,8 @@
 // ArduinoJson - arduinojson.org
-// Copyright Benoit Blanchon 2014-2020
+// Copyright Benoit Blanchon 2014-2019
 // MIT License
 
 #pragma once
-
-#include <ArduinoJson/Polyfills/pgmspace.hpp>
 
 namespace ARDUINOJSON_NAMESPACE {
 
@@ -16,7 +14,7 @@ class FlashStringAdapter {
     if (!other && !_str) return 0;
     if (!_str) return -1;
     if (!other) return 1;
-    return int8_t(-strcmp_P(other, reinterpret_cast<const char*>(_str)));
+    return -strcmp_P(other, reinterpret_cast<const char*>(_str));
   }
 
   bool equals(const char* expected) const {
